Refactoring
Separate functions into separate files, add import statements to primary index.ts file (now primarily handles endpoints only). Update HTML to add start button and add appropriate endpoint.
