+++ src/controllers/randomInclusiveNum.ts
@@ -1,0 +1,6 @@
+// Function to get a random integer between two values (inclusive)
+export function getRandomIntInclusive(min: number, max: number): number {
+  const minCeiled = Math.ceil(min);
+  const maxFloored = Math.floor(max);
+  return Math.floor(Math.random() * (maxFloored - minCeiled + 1) + minCeiled);
+}
